+++ Firebase/DynamicLinks/FIRDynamicLinks.m
@@ -194,8 +194,6 @@ static const NSInteger FIRErrorCodeDurableDeepLinkFailed = -119;
                                                        reason:errorDescription];
   }
   if (error) {
-    // TODO: Remove this call to sendLogsWithServiceName:version:.
-    [app sendLogsWithServiceName:@"DynamicLinks" version:kFIRDLVersion error:error];
     NSString *message = nil;
     if (options.usingOptionsFromDefaultPlist) {
       // Configured using plist file