+++ Firebase/InstanceID/FIRInstanceID.m
@@ -697,14 +697,7 @@ static FIRInstanceID *gInstanceID;
                          userInfo:userInfo];
 }
 
-// If the firebaseApp is available we should send logs for the error through it before
-// raising an exception.
 + (void)exitWithReason:(nonnull NSString *)reason forFirebaseApp:(FIRApp *)firebaseApp {
-  // TODO: Remove this call to sendLogsWithServiceName:version:.
-  [firebaseApp sendLogsWithServiceName:kFIRIIDServiceInstanceID
-                               version:FIRInstanceIDCurrentLibraryVersion()
-                                 error:[self configureErrorWithReason:reason]];
-
   [NSException raise:kFIRIIDErrorDomain
               format:@"Could not configure Firebase InstanceID. %@", reason];
 }